Added draft for a new GNSS-message containing ECEF
Added draft for a new GNSS-message containing ECEF coordinate representation: 1063.Fix2.uavcan.
In addition, a new field for a 3x3 covariance matrix is included:
pos_vel_covariance.
This describes the covariance between position and velocity.
Since covariance matrices are symmetric, only the upper-right quadrant
of the 6-by-6 matrix is required.