DTLS support in the experimental UDP forwarding type
Users shouldn't have to roll their own crypto. UAVCAN makes an implicit assumption that the intra-vehicular network is secure; being situated on the edge between the secure vehicular network and the outside world, modems can be burdened with security enforcement activities. Any UDP traffic can be tunneled via DTLS (it's one of the DTLS's design goals) virtually without changes to the application layer, so this type can be used as the optimal point of abstraction above the two.