Adopt a consistent power sign conversion.
Quote from Slack:

I think it removes one independent variable from our models: if you are dealing with a power source (a battery), you assume that positive current means that it is flowing into the network, and with ESC it's the opposite. If we adopt the power supply network model with the sign convention, then whether your entity is a battery or an ESC becomes irrelevant because its energy transfer relation to the power supply network becomes part of the model. It makes pieces more consistent with each other.