Remove the standard ESC registers.
This is because the fixed register names implicitly introduce the hard constraint that there may be at most one ESC per node.
This is not a reasonable restriction. Seeing as the MVP of DS-015 is not intended to support autoconfiguration,
it makes sense to simply omit this issue for now. It may be revised in v1.1 or later in a fashion that does not make it impossible
to run multiple service instances on the same node.