Break out Health and Mode from the Heartbeat due to the implications of the new alignment rules.
Keeping the fields unaligned might have prevented us from supporting enumerations later when they are introduced
because the new serialization model necessarily requires that all nested types are byte-aligned.

Another reason is that the old design (unchanged since the very first days of v0) violates the principles of
composability that were formulated relatively recently. For example, as the ongoing work on DS-015 indicated,
it is desirable to apply the concept of abstract component health to domain-specific services.