Refactor battery errors
Dmitry suggested that we add a few extra flags. That pushes the number of flags beyond the critical point where it no longer makes
sense to have them as separate states because excepting a single pair they are all mutually exclusive. In this changeset I removed
the FaultFlags and introduced the typical Error type that is a simple enumeration of errors that can be easily extended in the
future as necessary.