Provide explicit @extent everywhere to silence the diagnostics from PyDSDL
About the diagnostic: https://forum.uavcan.org/t/data-type-extensibility-and-composition/827/20?u=pavel.kirienko
...which was implemented in: https://github.com/UAVCAN/pydsdl/pull/45/commits/e2816c746fb3c00f4089cb97da00ff416626433d

Some of the values are rounded down to (integer power of two) - 32 to optimize dynamic memory allocations for
half-fit or buddy allocators which are expected to be commonly used in relevant applications. Other than that,
the extent values are more or less the best guess rather than a rigidly defined value.